requirements.txt
torch>=2.0.0
torchvision
numpy
flask
Pillow
scikit-learn
matplotlib
tqdm
